requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
mtcnn
Pillow